server/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
bcrypt
python-jose[cryptography]
fastapi-mail
python-dotenv
python-multipart
